Refactor all the code
It's easier to comprehend and work with when each block is distinct.
So I added a class to warning-block represents the input it is related
to and in JS created variables for every input that needs validation
and corresponding warning block. Then added eventListener for each of
the inputs and created its own validating functions.
Thus I split the logic and in the future if new inputs or validations
will appear it won't become a problem to add function and connect it.